Rename modules to avoid import namespace conflict
