+++ netsuite_bbc/constants.py
