+++ netsuite_bbc/util.py
